flaskapp-docker/requirements.txt
Pip==20.3.3
Pillow==7.2.0
Flask==1.1.2
Pytest==6.2.1
requests==2.25.1
PyPDF2==1.26.0
boto3==1.16.56
botocore==1.19.59   
jmespath==0.10.0  
python-dateutil==2.8.1   
s3transfer==0.3.4 
redis
flask-swagger-ui==3.36.0
